+++ lib/rubocop/path_util.rb
@@ -5,10 +5,14 @@ module RuboCop
   module PathUtil
     module_function
 
-    def relative_path(path, base_dir = Dir.pwd)
+    def relative_path(path, base_dir = PathUtil.pwd)
       # Optimization for the common case where path begins with the base
       # dir. Just cut off the first part.
-      return path[(base_dir.length + 1)..-1] if path.start_with?(base_dir)
+      if path.start_with?(base_dir)
+        base_dir_length = base_dir.length
+        result_length = path.length - base_dir_length - 1
+        return path[base_dir_length + 1, result_length]
+      end
 
       path_name = Pathname.new(File.expand_path(path))
       path_name.relative_path_from(Pathname.new(base_dir)).to_s
@@ -16,7 +20,7 @@ module RuboCop
 
     def smart_path(path)
       # Ideally, we calculate this relative to the project root.
-      base_dir = Dir.pwd
+      base_dir = PathUtil.pwd
 
       if path.start_with? base_dir
         relative_path(path, base_dir)
@@ -42,6 +46,14 @@ module RuboCop
     # Returns true for an absolute Unix or Windows path.
     def absolute?(path)
       path =~ %r{\A([A-Z]:)?/}
+    end
+
+    def self.pwd
+      @pwd ||= Dir.pwd
+    end
+
+    def self.reset_pwd
+      @pwd = nil
     end
   end
 end