+++ spec/spec_helper.rb
@@ -48,4 +48,8 @@ RSpec.configure do |config|
     mocks.syntax = :expect # Disable `should_receive` and `stub`
     mocks.verify_partial_doubles = true
   end
+
+  config.after do
+    RuboCop::PathUtil.reset_pwd
+  end
 end